added S3localstackActions
